Connected TodoList to Redux Store
